+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/GeckoSession.java
@@ -50,6 +50,7 @@ import java.io.InputStream;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.ref.WeakReference;
+import java.security.Principal;
 import java.security.cert.CertificateException;
 import java.security.cert.CertificateFactory;
 import java.security.cert.X509Certificate;
@@ -5952,10 +5953,26 @@ public class GeckoSession {
       /** The host requesting the certificate. */
       public final @NonNull String host;
 
+      /** The X.500 Distinguished Names the server specified as acceptable issuers. */
+      public final @Nullable Principal[] issuers;
+
+      @Deprecated
+      @DeprecationSchedule(id = "CertificateRequest-CertificateRequest", version = 143)
       protected CertificateRequest(
           final @NonNull String id, final Observer observer, final String host) {
         super(id, null, observer);
         this.host = host;
+        this.issuers = null;
+      }
+
+      protected CertificateRequest(
+          final @NonNull String id,
+          final Observer observer,
+          final String host,
+          final Principal[] issuers) {
+        super(id, null, observer);
+        this.host = host;
+        this.issuers = issuers;
       }
 
       /**