+++ security/manager/ssl/tests/mochitest/browser/browser_clientAuthRememberService.js
@@ -138,7 +138,7 @@ const gClientAuthDialogService = {
     this._chooseCertificateCalled = value;
   },
 
-  chooseCertificate(hostname, certArray, loadContext, callback) {
+  chooseCertificate(hostname, certArray, loadContext, caNames, callback) {
     this.chooseCertificateCalled = true;
     callback.certificateChosen(certArray[0], true);
   },